Updated sk to 2.05
darcs-hash:20060817085246-f1522-e0c77eae70575a6ff7552ff25c4d44d257841da2.gz

+++ src/filesystems/sleuthkit/auxtools/mymalloc.h
@@ -19,8 +19,8 @@ extern "C" {
     /*
      * External interface.
      */
-    extern char *mymalloc(int);
-    extern char *myrealloc(char *, int);
+    extern char *mymalloc(size_t);
+    extern char *myrealloc(char *, size_t);
     extern char *mystrdup(const char *);
 
 #ifdef __cplusplus
